CC3D - Add sonar.
This also ensures that the PWM mapping does not use the sonar pins when
sonar is enabled in a board agnostic way.
Conflicts:
	src/main/config/config.c
	src/main/drivers/pwm_mapping.h
	src/main/main.c
	src/main/target/CC3D/target.h

+++ src/main/drivers/timer.h
@@ -60,7 +60,7 @@ typedef struct timerOvrHandlerRec_s {
 typedef struct {
     TIM_TypeDef *tim;
     GPIO_TypeDef *gpio;
-    uint32_t pin;
+    uint16_t pin;
     uint8_t channel;
     uint8_t irq;
     uint8_t outputEnable;
